Add a pile of command-line options.

+++ streams/misc/release.sh
@@ -2,7 +2,7 @@
 #
 # release.sh: assemble a StreamIt release
 # David Maze <dmaze@cag.lcs.mit.edu>
-# $Id: release.sh,v 1.2 2003-03-10 17:13:35 dmaze Exp $
+# $Id: release.sh,v 1.3 2003-03-10 17:26:59 dmaze Exp $
 #
 
 # Interesting/configurable variables:
@@ -10,9 +10,49 @@ VERSION=0.0.20030314
 TAG=HEAD
 test -f /usr/share/java/antlrall.jar && ANTLRJAR=/usr/share/java/antlrall.jar
 test -f /usr/uns/java/antlr.jar && ANTLRJAR=/usr/uns/java/antlr.jar
+test -z "$TMPDIR" && TMPDIR=/tmp
+PRECIOUS=
+
+usage() {
+  cat >&2 <<EOF
+release.sh: assemble a StreamIt release
+
+Usage:
+  release.sh [options]
+
+Options:
+  --version (-v)   Use a particular exported version ($VERSION)
+  --tag (-t)       Build a release from a CVS tag ($TAG)
+  --cvsroot (-d)   Specify the CVS root directory ($CVSROOT)
+  --tmpdir         Use a different build directory ($TMPDIR)
+  --antlr          Location of the ANTLR jar file ($ANTLRJAR)
+  --precious (-k)  Keep the working directory
+EOF
+}
+
+# Command-line options:
+while test -n "$1"
+do
+  OPT="$1"; shift
+  case $OPT in
+    --version|-v) VERSION="$1"; shift;;
+    --tag|-t) TAG="$1"; shift;;
+    --tmpdir) TMPDIR="$1"; shift;;
+    --antlr) ANTLRJAR="$1"; shift;;
+    --cvsroot|-d) CVSROOT="$1"; export CVSROOT; shift;;
+    --precious|-k) PRECIOUS=yes;;
+    *) usage; exit 1;;
+  esac
+done
+
+# Make sure we have ANTLR somewhere.
+if test -z "$ANTLRJAR"; then
+  echo No ANTLR jar file\; use --antlr command-line option >&2
+  exit 1
+fi
 
 # Temporary directory:
-WORKING=${TMPDIR:=/tmp}/streamit-$USER-$$
+WORKING=$TMPDIR/streamit-$USER-$$
 mkdir $WORKING
 SRCDIR=$WORKING/streamit-$VERSION
 SRCTAR=$SRCDIR.tar
@@ -56,4 +96,10 @@ gzip $SRCTAR
 mv $SRCTAR.gz .
 
 # Clean up.
-rm -rf $WORKING
+if test -n "$PRECIOUS"
+then
+  echo Keeping working directory $WORKING
+else
+  rm -rf $WORKING
+fi
+
